tweaked cargo feature tests, think ctor was causing segfaults in vm
